refactor: rename test utils
