+++ crates/common/Cargo.toml
@@ -33,4 +33,4 @@ bip39 = "2.0.0"
 env_logger = { version = "0.11.0", default-features = false }
 test-log = { version = "0.2.12", default-features = false }
 wiremock = "0.5.19"
-test-tap-utils = { path = "../tests/tap-utils" }
+test-assets = { path = "../test-assets" }
