+++ crates/service/src/service/tap_receipt_header.rs
@@ -71,7 +71,7 @@ mod test {
     use axum::http::HeaderValue;
     use axum_extra::headers::Header;
 
-    use test_tap_utils::create_signed_receipt;
+    use test_assets::create_signed_receipt;
 
     use super::TapReceipt;
