+++ crates/service/src/tap/checks/deny_list_check.rs
@@ -202,7 +202,7 @@ mod tests {
     use tap_core::receipt::{Context, ReceiptWithState};
     use tokio::sync::watch;
 
-    use test_tap_utils::{
+    use test_assets::{
         self, create_signed_receipt, ESCROW_ACCOUNTS_BALANCES, ESCROW_ACCOUNTS_SENDERS_TO_SIGNERS,
         TAP_EIP712_DOMAIN, TAP_SENDER,
     };
